requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --no-emit-find-links --output-file=requirements/base.txt requirements/base.in
#
asgiref==3.8.1
    # via
    #   -r requirements/base.in
    #   django
    #   django-allauth
    #   django-browser-reload
    #   django-cors-headers
    #   django-htmx
authlib==1.5.1
    # via -r requirements/base.in
certifi==2025.4.26
    # via
    #   -r requirements/base.in
    #   sentry-sdk
cffi==1.17.1
    # via
    #   -r requirements/base.in
    #   cryptography
click==8.1.8
    # via -r requirements/base.in
contourpy==1.3.2
    # via
    #   -r requirements/base.in
    #   matplotlib
cryptography==41.0.7
    # via
    #   -r requirements/base.in
    #   authlib
    #   fido2
    #   python-dotenv-vault
cycler==0.12.1
    # via
    #   -r requirements/base.in
    #   matplotlib
django==5.2
    # via
    #   -r requirements/base.in
    #   django-allauth
    #   django-appconf
    #   django-browser-reload
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-htmx
    #   django-hyperscript
    #   django-pb-model
    #   django-select2
    #   django-tailwind
    #   django-template-partials
django-allauth==65.4.1
    # via -r requirements/base.in
django-appconf==1.1.0
    # via
    #   -r requirements/base.in
    #   django-select2
django-browser-reload==1.18.0
    # via -r requirements/base.in
django-cors-headers==4.7.0
    # via -r requirements/base.in
django-debug-toolbar==5.1.0
    # via -r requirements/base.in
django-htmx==1.22.0
    # via -r requirements/base.in
django-hyperscript==1.0.2
    # via -r requirements/base.in
django-matplotlib==0.1
    # via -r requirements/base.in
django-pb-model==0.3.3
    # via -r requirements/base.in
django-pdb==0.6.2
    # via -r requirements/base.in
django-select2==8.2.3
    # via -r requirements/base.in
django-tailwind==4.0.1
    # via -r requirements/base.in
django-template-partials==24.4
    # via -r requirements/base.in
fido2==1.2.0
    # via -r requirements/base.in
fonttools==4.57.0
    # via
    #   -r requirements/base.in
    #   matplotlib
kiwisolver==1.4.8
    # via
    #   -r requirements/base.in
    #   matplotlib
matplotlib==3.10.1
    # via -r requirements/base.in
numpy==2.2.5
    # via
    #   -r requirements/base.in
    #   contourpy
    #   matplotlib
packaging==25.0
    # via matplotlib
pillow==11.1.0
    # via
    #   -r requirements/base.in
    #   matplotlib
protobuf==6.30.2
    # via django-pb-model
pycparser==2.22
    # via cffi
pyparsing==3.2.3
    # via
    #   -r requirements/base.in
    #   matplotlib
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.in
    #   matplotlib
python-dotenv==0.21.1
    # via
    #   -r requirements/base.in
    #   python-dotenv-vault
python-dotenv-vault==0.6.4
    # via -r requirements/base.in
qrcode==8.0
    # via -r requirements/base.in
sentry-sdk==2.23.1
    # via -r requirements/base.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements/base.in
    #   django
    #   django-debug-toolbar
urllib3==2.4.0
    # via
    #   -r requirements/base.in
    #   sentry-sdk
